Move functionality to separate files
