+++ next/components/forms/segments/AccountSections/TaxesFeesSection/TaxesFeesSection.tsx
@@ -4,8 +4,7 @@ import TaxesFeesCard from 'components/forms/segments/AccountSections/TaxesFeesSe
 import TaxesFeesErrorCard from 'components/forms/segments/AccountSections/TaxesFeesSection/TaxesFeesErrorCard'
 // import TaxesFeesWaitingCard from 'components/forms/segments/AccountSections/TaxesFeesSection/TaxesFeesWaitingCard'
 import Spinner from 'components/forms/simple-components/Spinner'
-import { useDerivedServerSideAuthState, useTier } from 'frontend/hooks/useServerSideAuth'
-import { tierIdentityVerified } from 'frontend/utils/amplify'
+import { useDerivedServerSideAuthState } from 'frontend/hooks/useServerSideAuth'
 import { useTranslation } from 'next-i18next'
 
 import { ROUTES } from '../../../../../frontend/api/constants'
@@ -30,7 +29,6 @@ interface TaxesFeesSectionProps {
 const TaxesFeesSection: React.FC<TaxesFeesSectionProps> = () => {
   const { t } = useTranslation('account')
   const { tierStatus } = useDerivedServerSideAuthState()
-  // todo continue here tier comparisons
   const { data, isLoading } = useTaxes()
 
   //   const taxesFeesWaitingCardContent = `
