+++ next/components/forms/segments/CorrespondenceAddressForm/CorrespondenceAddressForm.tsx
@@ -1,7 +1,7 @@
 import AccountErrorAlert from 'components/forms/segments/AccountErrorAlert/AccountErrorAlert'
 import Button from 'components/forms/simple-components/Button'
 import InputField from 'components/forms/widget-components/InputField/InputField'
-import { AccountError, Address } from 'frontend/utils/amplify'
+import { AccountError, Address } from 'frontend/dtos/accountDto'
 import { useTranslation } from 'next-i18next'
 import { Controller } from 'react-hook-form'
