+++ next/components/forms/segments/CorrespondenceAddressModal/CorrespondenceAddressModal.tsx
@@ -1,6 +1,6 @@
 import CorrespondenceAddressForm from 'components/forms/segments/CorrespondenceAddressForm/CorrespondenceAddressForm'
 import Modal from 'components/forms/widget-components/Modals/Modal'
-import { AccountError, Address } from 'frontend/utils/amplify'
+import { AccountError, Address } from 'frontend/dtos/accountDto'
 import { useTranslation } from 'next-i18next'
 
 /* eslint-disable @typescript-eslint/no-shadow */
