+++ next/components/forms/segments/IdentityVerificationForm/IdentityVerificationForm.tsx
@@ -2,7 +2,7 @@ import ArrowRightIcon from '@assets/images/new-icons/ui/arrow-right.svg'
 import AccountErrorAlert from 'components/forms/segments/AccountErrorAlert/AccountErrorAlert'
 import Button from 'components/forms/simple-components/Button'
 import InputField from 'components/forms/widget-components/InputField/InputField'
-import { AccountError } from 'frontend/utils/amplify'
+import { AccountError } from 'frontend/dtos/accountDto'
 import { useRouter } from 'next/router'
 import { useTranslation } from 'next-i18next'
 import { useState } from 'react'
