+++ next/components/forms/segments/LoginForm/LoginForm.tsx
@@ -4,8 +4,8 @@ import Button from 'components/forms/simple-components/Button'
 import InputField from 'components/forms/widget-components/InputField/InputField'
 import PasswordField from 'components/forms/widget-components/PasswordField/PasswordField'
 import { ROUTES } from 'frontend/api/constants'
+import { AccountError } from 'frontend/dtos/accountDto'
 import useHookForm from 'frontend/hooks/useHookForm'
-import { AccountError } from 'frontend/utils/amplify'
 import { useTranslation } from 'next-i18next'
 import { Controller } from 'react-hook-form'
