+++ next/components/forms/segments/MigrationForm/MigrationForm.tsx
@@ -4,8 +4,8 @@ import LoginAccountLink from 'components/forms/segments/LoginAccountLink/LoginAc
 import Button from 'components/forms/simple-components/Button'
 import InputField from 'components/forms/widget-components/InputField/InputField'
 import { ROUTES } from 'frontend/api/constants'
+import { AccountError } from 'frontend/dtos/accountDto'
 import useHookForm from 'frontend/hooks/useHookForm'
-import { AccountError } from 'frontend/utils/amplify'
 import { formatUnicorn } from 'frontend/utils/string'
 import { useRouter } from 'next/router'
 import { useTranslation } from 'next-i18next'
