+++ next/components/forms/segments/PasswordChangeForm/PasswordChangeForm.tsx
@@ -1,8 +1,8 @@
 import AccountErrorAlert from 'components/forms/segments/AccountErrorAlert/AccountErrorAlert'
 import Button from 'components/forms/simple-components/Button'
 import PasswordField from 'components/forms/widget-components/PasswordField/PasswordField'
+import { AccountError } from 'frontend/dtos/accountDto'
 import useHookForm from 'frontend/hooks/useHookForm'
-import { AccountError } from 'frontend/utils/amplify'
 import { useTranslation } from 'next-i18next'
 import { Controller } from 'react-hook-form'
