+++ next/components/forms/segments/RegisterForm/RegisterForm.tsx
@@ -6,8 +6,8 @@ import InputField from 'components/forms/widget-components/InputField/InputField
 import PasswordField from 'components/forms/widget-components/PasswordField/PasswordField'
 import Radio from 'components/forms/widget-components/RadioButton/Radio'
 import RadioGroup from 'components/forms/widget-components/RadioButton/RadioGroup'
+import { AccountError, UserData } from 'frontend/dtos/accountDto'
 import useHookForm from 'frontend/hooks/useHookForm'
-import { AccountError, UserData } from 'frontend/utils/amplify'
 import { isBrowser } from 'frontend/utils/general'
 import logger from 'frontend/utils/logger'
 import { useTranslation } from 'next-i18next'
