+++ next/components/forms/segments/UserProfile/UserProfileView.tsx
@@ -1,10 +1,10 @@
 import { Auth } from 'aws-amplify'
 import cx from 'classnames'
 import MessageModal from 'components/forms/widget-components/Modals/MessageModal'
-import { usePageWrapperContext } from 'components/layouts/PageWrapper'
+import { AccountError, UserData } from 'frontend/dtos/accountDto'
+import { useRefreshServerSideProps } from 'frontend/hooks/useRefreshServerSideProps'
 import { useServerSideAuth } from 'frontend/hooks/useServerSideAuth'
 import useSnackbar from 'frontend/hooks/useSnackbar'
-import { AccountError, UserData, useRefreshServerSideProps } from 'frontend/utils/amplify'
 import logger from 'frontend/utils/logger'
 import { identity, mapValues, pickBy } from 'lodash'
 import { useTranslation } from 'next-i18next'
