+++ next/pages/logout.tsx
@@ -2,8 +2,8 @@ import { Auth } from 'aws-amplify'
 import AccountContainer from 'components/forms/segments/AccountContainer/AccountContainer'
 import AccountSuccessAlert from 'components/forms/segments/AccountSuccessAlert/AccountSuccessAlert'
 import LoginRegisterLayout from 'components/layouts/LoginRegisterLayout'
+import { getSSRCurrentAuth } from 'components/logic/ServerSideAuthProvider'
 import useSSORedirect from 'frontend/hooks/useSSORedirect'
-import { getSSRCurrentAuth } from 'frontend/utils/amplify'
 import logger from 'frontend/utils/logger'
 import { GetServerSidePropsContext } from 'next'
 import { useTranslation } from 'next-i18next'
