+++ next/pages/moje-ziadosti/[ziadost].tsx
@@ -1,8 +1,8 @@
 import MyApplicationDetails from 'components/forms/segments/AccountSections/MyApplicationsSection/MyApplicationDetails'
 import AccountPageLayout from 'components/layouts/AccountPageLayout'
 import PageWrapper from 'components/layouts/PageWrapper'
+import { getSSRCurrentAuth } from 'components/logic/ServerSideAuthProvider'
 import { getApplicationDetailsData, getApplicationHistoryData } from 'frontend/api/mocks/mocks'
-import { getSSRCurrentAuth } from 'frontend/utils/amplify'
 import { isProductionDeployment } from 'frontend/utils/general'
 import logger from 'frontend/utils/logger'
 import { AsyncServerProps } from 'frontend/utils/types'
