+++ next/pages/moje-ziadosti/index.tsx
@@ -1,8 +1,8 @@
 import MyApplicationsSection from 'components/forms/segments/AccountSections/MyApplicationsSection/MyApplicationsSection'
 import AccountPageLayout from 'components/layouts/AccountPageLayout'
 import PageWrapper from 'components/layouts/PageWrapper'
+import { getSSRCurrentAuth } from 'components/logic/ServerSideAuthProvider'
 import { getApplicationConceptList, getApplicationSentList } from 'frontend/api/mocks/mocks'
-import { getSSRCurrentAuth } from 'frontend/utils/amplify'
 import { isProductionDeployment } from 'frontend/utils/general'
 import logger from 'frontend/utils/logger'
 import { AsyncServerProps } from 'frontend/utils/types'
