+++ next/pages/registracia.tsx
@@ -6,8 +6,9 @@ import AccountSuccessAlert from 'components/forms/segments/AccountSuccessAlert/A
 import EmailVerificationForm from 'components/forms/segments/EmailVerificationForm/EmailVerificationForm'
 import RegisterForm from 'components/forms/segments/RegisterForm/RegisterForm'
 import LoginRegisterLayout from 'components/layouts/LoginRegisterLayout'
+import { getSSRCurrentAuth } from 'components/logic/ServerSideAuthProvider'
+import { AccountError, UserData } from 'frontend/dtos/accountDto'
 import useSSORedirect from 'frontend/hooks/useSSORedirect'
-import { AccountError, AccountStatus, getSSRCurrentAuth, UserData } from 'frontend/utils/amplify'
 import { GetServerSidePropsContext } from 'next'
 import { useRouter } from 'next/router'
 import { useTranslation } from 'next-i18next'
