+++ next/pages/styleguide.tsx
@@ -1,5 +1,6 @@
 import { StatusBar } from 'components/forms/info-components/StatusBar'
 import PageWrapper from 'components/layouts/PageWrapper'
+import { getSSRCurrentAuth } from 'components/logic/ServerSideAuthProvider'
 import DatePickerShowCase from 'components/styleguide/showcases/DatePickerShowCase'
 import InputFieldShowCase from 'components/styleguide/showcases/InputFieldShowCase'
 import TimePickerShowCase from 'components/styleguide/showcases/TimePickerShowCase'
